drivers: Trial and Error #3

+++ drivers/thermal/msm_thermal.c
@@ -2606,7 +2606,7 @@ set_done:
 
 int32_t tsens_get_temp(struct tsens_device *dev, long *temp);
 
-static int therm_get_temp(uint32_t id, enum sensor_id_type type, long *temp)
+static int therm_get_temp(uint32_t id, enum sensor_id_type type, int *temp)
 {
 	int ret = 0;
 	struct tsens_device tsens_dev;
